+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/crowdfundingbox/bytecode/fingerprints/CrowdfundingBoxFingerprint.kt
@@ -7,12 +7,9 @@ import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 
 object CrowdfundingBoxFingerprint : MethodFingerprint(
     opcodes = listOf(
-        Opcode.CONST_4,
-        Opcode.CONST,
         Opcode.INVOKE_VIRTUAL,
         Opcode.MOVE_RESULT_OBJECT,
         Opcode.IPUT_OBJECT,
-        Opcode.CONST,
     ),
     customFingerprint = { methodDef ->
         methodDef.implementation?.instructions?.any { instruction ->