+++ src/main/kotlin/app/revanced/patches/youtube/misc/fix/playback/patch/SpoofSignatureVerificationPatch.kt
@@ -19,7 +19,9 @@ import app.revanced.patches.shared.settings.preference.impl.SwitchPreference
 import app.revanced.patches.youtube.misc.fix.playback.annotation.ProtobufSpoofCompatibility
 import app.revanced.patches.youtube.misc.fix.playback.fingerprints.OpenCronetDataSourceFingerprint
 import app.revanced.patches.youtube.misc.fix.playback.fingerprints.ProtobufParameterBuilderFingerprint
+import app.revanced.patches.youtube.misc.fix.playback.fingerprints.SubtitleWindowSettingsConstructorFingerprint
 import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
+import app.revanced.patches.youtube.misc.playertype.patch.PlayerTypeHookPatch
 import app.revanced.patches.youtube.misc.settings.bytecode.patch.SettingsPatch
 import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
 
@@ -27,12 +29,13 @@ import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
 @Name("spoof-signature-verification")
 @Description("Spoofs the client to prevent playback issues.")
 @ProtobufSpoofCompatibility
-@DependsOn([IntegrationsPatch::class, SettingsPatch::class])
+@DependsOn([IntegrationsPatch::class, SettingsPatch::class, PlayerTypeHookPatch::class])
 @Version("0.0.1")
 class SpoofSignatureVerificationPatch : BytecodePatch(
     listOf(
         ProtobufParameterBuilderFingerprint,
         OpenCronetDataSourceFingerprint,
+        SubtitleWindowSettingsConstructorFingerprint,
     )
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
@@ -42,7 +45,9 @@ class SpoofSignatureVerificationPatch : BytecodePatch(
                 StringResource("revanced_spoof_signature_verification_title", "Spoof app signature"),
                 false,
                 StringResource("revanced_spoof_signature_verification_summary_on", "App signature spoofed"),
-                StringResource("revanced_spoof_signature_verification_summary_off", "App signature not spoofed")
+                StringResource("revanced_spoof_signature_verification_summary_off", "App signature not spoofed"),
+                StringResource("revanced_spoof_signature_verification_user_dialog_message",
+                    "Signature spoofing can fix playback issues, but may causes side effects.")
             )
         )
 
@@ -82,10 +87,29 @@ class SpoofSignatureVerificationPatch : BytecodePatch(
 
         } ?: return OpenCronetDataSourceFingerprint.toErrorResult()
 
+        // hook override subtitles
+        SubtitleWindowSettingsConstructorFingerprint.result?.let {
+            it.mutableMethod.apply {
+                addInstructions(
+                    0,
+                    """
+                        invoke-static {p1, p2, p3, p4, p5}, $INTEGRATIONS_CLASS_DESCRIPTOR->getSubtitleWindowSettingsOverride(IIIZZ)[I
+                        move-result-object v0
+                        const/4 v1, 0x0
+                        aget p1, v0, v1     # ap, anchor configuration
+                        const/4 v1, 0x1
+                        aget p2, v0, v1     # ah, horizontal anchor
+                        const/4 v1, 0x2
+                        aget p3, v0, v1     # av, vertical anchor
+                    """
+                )
+            }
+        } ?: return SubtitleWindowSettingsConstructorFingerprint.toErrorResult()
+
         return PatchResultSuccess()
     }
 
-    companion object {
+    private companion object {
         const val INTEGRATIONS_CLASS_DESCRIPTOR = "Lapp/revanced/integrations/patches/SpoofSignatureVerificationPatch;"
     }
 }