+++ src/main/kotlin/app/revanced/patches/youtube/misc/settings/resource/patch/SettingsResourcePatch.kt
@@ -18,6 +18,7 @@ import app.revanced.patches.shared.settings.resource.patch.AbstractSettingsResou
 import app.revanced.patches.youtube.misc.settings.bytecode.patch.SettingsPatch
 import app.revanced.util.resources.ResourceUtils
 import app.revanced.util.resources.ResourceUtils.copyResources
+import app.revanced.util.resources.ResourceUtils.mergeStrings
 import org.w3c.dom.Node
 
 @Name("settings-resource-patch")
@@ -76,6 +77,8 @@ class SettingsResourcePatch : AbstractSettingsResourcePatch(
                 StringResource("revanced_settings_summary", "ReVanced specific settings"),
             )
         )
+
+        context.mergeStrings("settings/host/values/strings.xml")
 
         return PatchResultSuccess()
     }